backend/requirements.txt
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
kombu==5.4.2
numpy==2.2.1
opencv-python-headless==4.10.0.84
pillow==11.1.0
psycopg[binary]==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
sqlparse==0.5.3
tzdata==2025.1
vine==5.1.0
